Extend entity-table with interactive controls in cells
Adapt vm-list to use new controls
