--- src/app/pages/services/services.component.html
@@ -1,1 +1,0 @@
-<services-table [conf]="this"></services-table>
